+++ web/src/components/Header.tsx
@@ -1,19 +1,15 @@
 import { useEffect } from "react";
 import { NavLink, useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import { useGlobalStore, useLayoutStore, useUserStore } from "@/store/module";
+import { useLayoutStore, useUserStore } from "@/store/module";
 import { resolution } from "@/utils/layout";
 import Icon from "./Icon";
 import UserBanner from "./UserBanner";
-import showSettingDialog from "./SettingDialog";
-import showAskAIDialog from "./AskAIDialog";
-import showArchivedMemoDialog from "./ArchivedMemoDialog";
 import showAboutSiteDialog from "./AboutSiteDialog";
 import showMemoEditorDialog from "./MemoEditor/MemoEditorDialog";
 
 const Header = () => {
   const { t } = useTranslation();
-  const globalStore = useGlobalStore();
   const location = useLocation();
   const userStore = useUserStore();
   const layoutStore = useLayoutStore();
@@ -109,37 +105,40 @@ const Header = () => {
           </NavLink>
           {!isVisitorMode && (
             <>
-              <button
-                id="header-ask-ai"
-                className="px-4 pr-5 py-2 rounded-full flex flex-row items-center text-lg text-gray-800 dark:text-gray-300 hover:bg-white hover:shadow dark:hover:bg-zinc-700"
-                onClick={() => showAskAIDialog()}
+              <NavLink
+                to="/archived"
+                id="header-setting"
+                className={({ isActive }) =>
+                  `${
+                    isActive && "bg-white dark:bg-zinc-700 shadow"
+                  } px-4 pr-5 py-2 rounded-full flex flex-row items-center text-lg text-gray-800 dark:text-gray-300 hover:bg-white hover:shadow dark:hover:bg-zinc-700`
+                }
               >
-                <Icon.Bot className="mr-3 w-6 h-auto opacity-70" /> {t("ask-ai.title")}
-              </button>
-              <button
-                id="header-archived-memo"
-                className="px-4 pr-5 py-2 rounded-full flex flex-row items-center text-lg text-gray-800 dark:text-gray-300 hover:bg-white hover:shadow dark:hover:bg-zinc-700"
-                onClick={() => showArchivedMemoDialog()}
+                <>
+                  <Icon.Archive className="mr-3 w-6 h-auto opacity-70" /> {t("common.archived")}
+                </>
+              </NavLink>
+              <NavLink
+                to="/setting"
+                id="header-setting"
+                className={({ isActive }) =>
+                  `${
+                    isActive && "bg-white dark:bg-zinc-700 shadow"
+                  } px-4 pr-5 py-2 rounded-full flex flex-row items-center text-lg text-gray-800 dark:text-gray-300 hover:bg-white hover:shadow dark:hover:bg-zinc-700`
+                }
               >
-                <Icon.Archive className="mr-3 w-6 h-auto opacity-70" /> {t("common.archived")}
-              </button>
-              <button
-                id="header-settings"
-                className="px-4 pr-5 py-2 rounded-full flex flex-row items-center text-lg text-gray-800 dark:text-gray-300 hover:bg-white hover:shadow dark:hover:bg-zinc-700"
-                onClick={() => showSettingDialog()}
-              >
-                <Icon.Settings className="mr-3 w-6 h-auto opacity-70" /> {t("common.settings")}
-              </button>
-              {globalStore.isDev() && (
-                <div className="pr-3 pl-1 w-full">
-                  <button
-                    className="mt-2 w-full py-3 rounded-full flex flex-row justify-center items-center bg-green-600 font-medium text-white dark:opacity-80 hover:shadow hover:opacity-90"
-                    onClick={() => showMemoEditorDialog()}
-                  >
-                    <Icon.Edit3 className="w-4 h-auto mr-1" /> New
-                  </button>
-                </div>
-              )}
+                <>
+                  <Icon.Settings className="mr-3 w-6 h-auto opacity-70" /> {t("common.settings")}
+                </>
+              </NavLink>
+              <div className="pr-3 pl-1 w-full">
+                <button
+                  className="mt-2 w-full py-3 rounded-full flex flex-row justify-center items-center bg-green-600 font-medium text-white dark:opacity-80 hover:shadow hover:opacity-90"
+                  onClick={() => showMemoEditorDialog()}
+                >
+                  <Icon.Edit3 className="w-4 h-auto mr-1" /> New
+                </button>
+              </div>
             </>
           )}
           {isVisitorMode && (
