--- web/src/less/archived-memo-dialog.less
@@ -1,19 +1,0 @@
-.archived-memo-dialog {
-  @apply px-4;
-
-  > .dialog-container {
-    @apply w-128 max-w-full mb-8;
-
-    > .dialog-content-container {
-      @apply w-full flex flex-col justify-start items-start;
-
-      > .tip-text-container {
-        @apply w-full h-32 flex flex-col justify-center items-center;
-      }
-
-      > .archived-memos-container {
-        @apply w-full flex flex-col justify-start items-start;
-      }
-    }
-  }
-}
