+++ web/src/less/archived.less
@@ -1,0 +1,11 @@
+.archived-memo-page {
+  @apply w-full max-w-3xl min-h-full flex flex-col justify-start items-center pb-8 bg-zinc-100 dark:bg-zinc-800;
+
+  > .tip-text-container {
+    @apply w-full h-32 flex flex-col justify-center items-center;
+  }
+
+  > .archived-memos-container {
+    @apply w-full flex flex-col justify-start items-start;
+  }
+}
