+++ web/src/less/memo-list.less
@@ -1,5 +1,5 @@
 .memo-list-container {
-  @apply flex flex-col justify-start items-start w-full max-w-full overflow-y-scroll pb-16 hide-scrollbar;
+  @apply flex flex-col justify-start items-start w-full max-w-full overflow-y-scroll pb-28 hide-scrollbar;
 
   > .status-text-container {
     @apply flex flex-col justify-start items-center w-full my-6;
