--- web/src/less/setting-dialog.less
@@ -1,59 +1,0 @@
-.setting-dialog {
-  @apply px-4;
-
-  > .dialog-container {
-    @apply w-180 max-w-full h-full sm:h-auto mb-8 p-0;
-
-    > .dialog-content-container {
-      @apply flex flex-row justify-start items-start relative w-full h-full p-0;
-
-      > .close-btn {
-        @apply z-1 flex flex-col justify-center items-center absolute top-4 right-4 w-6 h-6 rounded hover:bg-gray-200 dark:hover:bg-zinc-700 hover:shadow;
-      }
-
-      > .section-selector-container {
-        @apply hidden sm:flex flex-col justify-start items-start sm:w-52 h-auto sm:h-full shrink-0 rounded-t-lg sm:rounded-none sm:rounded-l-lg p-4 bg-gray-100 dark:bg-zinc-700;
-
-        > .section-title {
-          @apply text-sm mt-2 sm:mt-4 first:mt-4 mb-1 font-mono text-gray-400;
-        }
-
-        > .section-items-container {
-          @apply w-full h-auto flex flex-row sm:flex-col justify-start items-start;
-
-          > .section-item {
-            @apply flex flex-row justify-start items-center text-base select-none mr-3 sm:mr-0 mt-0 sm:mt-2 text-gray-700 dark:text-gray-300 cursor-pointer hover:opacity-80;
-
-            &.selected {
-              @apply font-bold hover:opacity-100;
-            }
-
-            > .icon-text {
-              @apply text-base mr-2;
-            }
-          }
-        }
-      }
-
-      > .section-content-container {
-        @apply w-full sm:w-auto p-4 sm:px-6 grow flex flex-col justify-start items-start h-full sm:h-128 overflow-y-scroll hide-scrollbar;
-
-        > .section-container {
-          @apply flex flex-col justify-start items-start w-full my-2;
-
-          .title-text {
-            @apply text-sm mt-4 first:mt-2 mb-3 font-mono text-gray-500;
-          }
-
-          > .form-label {
-            @apply flex flex-row items-center w-full mb-2;
-
-            > .normal-text {
-              @apply shrink-0 select-text;
-            }
-          }
-        }
-      }
-    }
-  }
-}
