+++ web/src/pages/Home.tsx
@@ -32,7 +32,7 @@ function Home() {
 
   return (
     <div className="w-full flex flex-row justify-start items-start">
-      <div className="flex-grow shrink w-auto max-w-2xl px-4 sm:px-2 sm:pt-4">
+      <div className="flex-grow shrink w-auto px-4 sm:px-2 sm:pt-4">
         <MobileHeader />
         <div className="w-full h-auto flex flex-col justify-start items-start bg-zinc-100 dark:bg-zinc-800 rounded-lg">
           {!userStore.isVisitorMode() && <MemoEditor />}
