+++ web/src/pages/ResourcesDashboard.tsx
@@ -194,7 +194,7 @@ const ResourcesDashboard = () => {
   };
 
   return (
-    <section className="w-full max-w-2xl min-h-full flex flex-col justify-start items-center px-4 sm:px-2 sm:pt-4 pb-8 bg-zinc-100 dark:bg-zinc-800">
+    <section className="w-full max-w-3xl min-h-full flex flex-col justify-start items-center px-4 sm:px-2 sm:pt-4 pb-8 bg-zinc-100 dark:bg-zinc-800">
       <MobileHeader showSearch={false} />
       <div className="w-full relative" onDragEnter={handleDrag}>
         {dragActive && (
