+++ web/src/router/index.tsx
@@ -5,6 +5,8 @@ import store from "@/store";
 import { initialGlobalState, initialUserState } from "@/store/module";
 import DailyReview from "@/pages/DailyReview";
 import ResourcesDashboard from "@/pages/ResourcesDashboard";
+import Setting from "@/pages/Setting";
+import Archived from "@/pages/Archived";
 
 const Root = lazy(() => import("@/layouts/Root"));
 const Auth = lazy(() => import("@/pages/Auth"));
@@ -129,6 +131,44 @@ const router = createBrowserRouter([
       {
         path: "resources",
         element: <ResourcesDashboard />,
+        loader: async () => {
+          await initialGlobalStateLoader();
+
+          try {
+            await initialUserState();
+          } catch (error) {
+            // do nth
+          }
+
+          const { host } = store.getState().user;
+          if (isNullorUndefined(host)) {
+            return redirect("/auth");
+          }
+          return null;
+        },
+      },
+      {
+        path: "archived",
+        element: <Archived />,
+        loader: async () => {
+          await initialGlobalStateLoader();
+
+          try {
+            await initialUserState();
+          } catch (error) {
+            // do nth
+          }
+
+          const { host } = store.getState().user;
+          if (isNullorUndefined(host)) {
+            return redirect("/auth");
+          }
+          return null;
+        },
+      },
+      {
+        path: "setting",
+        element: <Setting />,
         loader: async () => {
           await initialGlobalStateLoader();
