+++ web/src/store/zustand/index.ts
@@ -1,2 +1,1 @@
 export { useMemoCacheStore } from "./memo";
-export { useMessageStore } from "./message";
